+++ packages/node/src/index.ts
@@ -3,5 +3,5 @@ export type { Queryable, VersionedUpdateQuery } from "./database.js";
 export { MomentNodeError } from "./errors.js";
 export type { MomentNodeErrorCode } from "./errors.js";
 export { assertStrongEtag, captureHttpCommitState, captureHttpObservation, conditionalHeaders, guardedFetch } from "./http.js";
-export { generateSigningKeyPair, keyIdFromPublicKey, signReceipt, verifySignedReceipt } from "./signature.js";
-export type { SignedReceiptEnvelope, SignedReceiptVerification } from "./signature.js";
+export { generateSigningKeyPair, keyIdFromPublicKey, receiptSigningPayload, signReceipt, signReceiptWithSigner, verifySignedReceipt, verifySignedReceiptWithTrustStore } from "./signature.js";
+export type { ReceiptSigner, SignedReceiptEnvelope, SignedReceiptVerification, TrustedReceiptKey } from "./signature.js";